// src/app/main/apps/procesos/contratacion/models/contratacion.model.ts
export class DocumentoSustentoModel {
    idDocumentoSustento: number = null;
    idLicencia: number = null;
    idTipoDocumentoSustento: number = null;
    idTipoFormatoSustento: number = null;
    numeroDocumentoSustento: string = null;
    entidadEmisora: string = null;
    fechaEmision: string = null;
    numeroFolios: number = null;
    sumilla: string = null;
    codigoDocumentoSustento: number = null;
    codigoAdjuntoSustento: string = null;
    fechaRegistro: string = null;
    descripcionTipoSustento: string;
    descripcionTipoFormato: string;
}
// ...
export class PostulacionModel {
    idPostulacion: number;
    idProceso: number;
    idEtapa: number;
    idPersona: number;
    idOrigenRegistro: number;
    idTipoVia: number;
    idTipoZona: number;
    idPais: number;
    idDistrito: number;
    idColegioProfesional?: number;
    idEstadoPostulacion: number;
    codigoPlaza: string;
    nombreVia: string;
    nombreZona: string;
    numeroVia: number;
    direccion: string;
    telefonoMovil: string;
    telefonoFijo: string;
    numeroRuc: string;
    numeroColegioProfesional: string;
    codigoPostulante: number;
    usuarioRegistro: string = null;
    ipRegistro: string = null;
    capacitaciones: CapacitacionModel[] = [];
    formacionesAcademica: FormacionAcademicaModel[] = [];
    experienciasLaborales: ExperienciaLaboralModel[] = [];
    datoPlazaPostular: DatoPlazaPostularModel = null;
    documentosSustento: DocumentoSustentoModel[] = [];

    idDepartamento: number = null;
    idProvincia: number = null;
    idTipoDocumentoIdentidad: number = null;
    numeroDocumentoIdentidad: string = null;
    numeroDocumentoIdentidadCompleto: string = null;
    fechaNacimiento: string = null;
    primerApellido: string = null;
    segundoApellido: string = null;
    nombres: string = null;
    descripcionDepartamento: string = null;
    descripcionProvincia: string = null;
    descripcionDistrito: string = null;
    descripcionGenero: string = null;
    descripcionNacionalidad: string = null;
    descripcionTipoVia: string = null;
    descripcionTipoZona: string = null;
    nombreColegioProfesional: string = null;

    experienciaLaboralCalc: ExperienciaLaboralCalc = null;
    plazaPostulada: PlazaPostuladaModel;
}

export class CapacitacionModel {
    idCapacitacion: number = null;
    idPostulacion?: number = null;
    idTipoCapacitacion: number = null;
    idPais: number = null;
    nombreCapacitacion: string = null;
    fechaInicio: Date = null;
    fechaFin: Date = null;
    duracionHoras: number = null;
    institucionCapacitacion: string = null;
    lugarCapacitacion: string = null;
    codigoAdjuntoSustento?: string = null;
    codigoCapacitacion: number = null;
    fechaCreacion: Date = null;
    descripcionTipoCapacitacion: string = null;
}
export class FormacionAcademicaModel {
    idFormacionAcademica: number = null;
    idPostulacion: number = null;
    idNivelEducativo: number = null;
    idGradoEstudio: number = null;
    idAreaCurricular: number = null;
    idCentroEstudio: number = null;
    idEstudiosCompletos: number = null;
    codigoFormacionAcademica: number = null;
    anioInicioEstudios: number = null;
    anioFinEstudios?: number = null;
    fechaExpedicionGradoEstudios: Date = null;
    codigoAdjuntoSustento?: string = null;
    beneficiarioPronabec?: boolean = null;
    fechaRegistro: Date = null;

    descripcionNivelEducativo: string = null;
    descripcionGradoEstudio: string = null;
    descripcionAreaCurricular: string = null;
    descripcionCentroEstudio: string = null;
    descripcionEstudiosCompletos: string = null;
}
export class ExperienciaLaboralModel {
    idExperienciaLaboral: number = null;
    idPostulacion: number = null;
    idTipoExperiencia: number = null;
    idNivelExperiencia: number = null;
    idTipoEntidad: number = null;
    cargoExperienciaLaboral: string = null;
    fechaInicio: Date = null;
    fechaFin: Date = null;
    nombreEntidad: string = null;
    codigoAdjuntoSustento?: string = null;
    codigoExperienciaLaboral: number = null;
    descripcionTipoExperiencia: string = null;
    descripcionNivelExperiencia: string = null;
    descripcionTipoEntidad: string = null;
}

export class DatoPlazaPostularModel {
    idDatoPostular: number = null;
    idCargo: number = null;
    idModalidadEducativa: number = null;
    idNivelEducativo: number = null;
    idAreaCurricular: number = null;
    descripcionCargo: string;
    descripcionModalidadEducativa: string;
    descripcionNivelEducativo: string;
    descripcionAreaCurricular: string;
}
// ...
export class ServidorPublicoModel {
    registro: number;
    totalRegistro: number;
    idServidorPublico: number;
    idPersona: number;
    idTipoDocumentoIdentidad: number;
    numeroDocumentoIdentidad: string;
    numeroDocumentoIdentidadCompleto: string;
    nombreCompleto: string;
    fechaNacimiento: string;
    descripcionUgel: string;
    centroTrabajo: string;
    abreviaturaRegimenLaboral: string;
    situacionLaboral: string;
    idSituacionLaboral: number;
    descripcionGenero: string;
    descripcionNacionalidad: string;
    primerApellido: string;
    segundoApellido: string;
    nombres: string;
}

export class ExperienciaLaboralCalc {
    totalAnioGeneral: number = null;
    totalmesGeneral: number = null;
    diasRestanteGeneral: number = null;
    totalAnioEspecifico: number = null;
    totalmesEspecifico: number = null;
    diasRestanteEspecifico: number = null;
}
// ...
export class PlazaPostuladaModel {
    idPlazaPostulada: number;
    idPlaza: number;
    codigoModular: string;
    descripcionInstitucionEducativa: string;
    abreviaturaModalidadEducativa: string;
    descripcionNivelEducativo: string;
    descripcionTipoGestionEducativa: string;
    nombreZona: string;
    eib: string;
    codigoPlaza: string;
    descripcionCargo: string;
    descripcionAreaCurricular: string;
    jornadaLaboral: string;
    descripcionTipoPlaza: string;
    descripcionMotivoVacancia: string;
    codigoCentroTrabajo: string;
    idPlazaContratacion: number;
    descripcionRegimenLaboral: string;
    idGrupoInscripcion: number;
    descripcionGrupoInscripcion: string;
}